Fixed Things Missed or Broken from Commit d47bd14

+++ Processing_Sample/CProcessing/Source/Internal_System.h
@@ -98,7 +98,7 @@ extern "C" {
 	CP_CorePtr GetCPCore(void);
 	CP_DrawInfoPtr GetDrawInfo(void);
 
-	void CP_SetWindowSizeInternal(int new_width, int new_height, bool isFullscreen);
+	void CP_SetWindowSizeInternal(unsigned short new_width, unsigned short new_height, bool isFullscreen);
 
 	void CP_Initialize(void);
 	void CP_Update(void);
